backend/requirements.txt
Django>=6.0.0
djangorestframework
django-cors-headers
drf-spectacular
psycopg2-binary
python-dotenv
